+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.2.0
+* Updated readme instructions 
+* "Group->Show category" and "Group->Font size" were not working
+* Add colors selection in settings panel
+* Add an ability to show values of the inner ring
+* Add an ability to support decimal formatting
+
 ## 1.1.1
 * Added multi language resource files